+++ webpack.config.js
@@ -1,87 +1,18 @@
 'use strict';
 
 const path = require('path');
-const webpack = require('webpack');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
 const proxy = require('./server/webpack-dev-proxy');
 const loaders = require('./webpack/loaders');
-const StyleLintPlugin = require('stylelint-webpack-plugin');
-const SplitByPathPlugin = require('webpack-split-by-path');
+const plugins = require('./webpack/plugins');
+const postcssInit = require('./webpack/postcss');
 
-const baseAppEntries = [
-  './src/index.tsx',
-];
-
-const devAppEntries = [
-  'webpack-hot-middleware/client?reload=true',
-];
-
+const baseAppEntries = [ './src/index.tsx' ];
+const devAppEntries = [ 'webpack-hot-middleware/client?reload=true' ];
 const appEntries = baseAppEntries
   .concat(process.env.NODE_ENV === 'development' ? devAppEntries : []);
 
-const basePlugins = [
-  new webpack.DefinePlugin({
-    __DEV__: process.env.NODE_ENV !== 'production',
-    'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
-  }),
-  new SplitByPathPlugin([
-    { name: 'vendor', path: [__dirname + '/node_modules/'] },
-  ]),
-  new HtmlWebpackPlugin({
-    template: './src/index.html',
-    inject: 'body',
-  }),
-];
-
-const devPlugins = [
-  new webpack.NoErrorsPlugin(),
-  new StyleLintPlugin({
-    configFile: './.stylelintrc',
-    files: ['src/**/*.css'],
-    failOnError: false,
-  }),
-];
-
-const prodPlugins = [
-  new webpack.optimize.UglifyJsPlugin({
-    compress: {
-      warnings: false,
-    },
-  }),
-  new webpack.NoErrorsPlugin(),
-];
-
-const plugins = basePlugins
-  .concat(process.env.NODE_ENV === 'production' ? prodPlugins : [])
-  .concat(process.env.NODE_ENV === 'development' ? devPlugins : []);
-
-const postcssBasePlugins = [
-  require('postcss-modules-local-by-default'),
-  require('postcss-import')({
-    addDependencyTo: webpack,
-  }),
-  require('postcss-cssnext')({
-    browsers: ['ie >= 8', 'last 2 versions'],
-  }),
-];
-const postcssDevPlugins = [];
-const postcssProdPlugins = [
-  require('cssnano')({
-    safe: true,
-    sourcemap: true,
-    autoprefixer: false,
-  }),
-];
-
-const postcssPlugins = postcssBasePlugins
-  .concat(process.env.NODE_ENV === 'production' ? postcssProdPlugins : [])
-  .concat(process.env.NODE_ENV === 'development' ? postcssDevPlugins : []);
-
 module.exports = {
-
-  entry: {
-    app: appEntries,
-  },
+  entry: { app: appEntries },
 
   output: {
     path: path.join(__dirname, 'dist'),
@@ -92,11 +23,7 @@ module.exports = {
   },
 
   devtool: 'source-map',
-
-  resolve: {
-    extensions: ['', '.webpack.js', '.web.js', '.tsx', '.ts', '.js'],
-  },
-
+  resolve: { extensions: ['', '.webpack.js', '.web.js', '.tsx', '.ts', '.js'] },
   plugins: plugins,
 
   devServer: {
@@ -120,7 +47,5 @@ module.exports = {
     ],
   },
 
-  postcss: () => {
-    return postcssPlugins;
-  },
+  postcss: postcssInit,
 };
